project/backend/requirements.txt
# Core dependencies
fastapi>=0.104.0,<0.115.0
uvicorn[standard]>=0.24.0,<0.32.0
python-multipart>=0.0.6,<0.1.0

# LangChain
langchain>=0.2.0,<0.4.0
langchain-community>=0.2.0,<0.4.0
langchain-postgres>=0.0.12,<0.1.0

# Document processing
pypdf>=4.0.0,<5.0.0
pytesseract>=0.3.10,<0.4.0
pdf2image>=1.16.0,<2.0.0
Pillow>=9.5.0,<11.0.0

# Database
psycopg2-binary>=2.9.0,<3.0.0
sqlalchemy>=2.0.0,<2.1.0
asyncpg>=0.29.0,<0.30.0

# Utilities
pydantic>=2.5.0,<3.0.0
python-dotenv>=1.0.0,<2.0.0
httpx>=0.25.0,<0.28.0

# ML dependencies (의존성 명시적 포함)
transformers>=4.30.0,<5.0.0
sentence-transformers>=2.2.0,<3.0.0
numpy>=1.24.0,<2.0.0
tqdm>=4.27
tokenizers>=0.19.0
huggingface-hub>=0.15.0